+++ src/clients/schemas.ts
@@ -40,6 +40,38 @@ const JobDetailsSchema = z.object({
   }),
 });
 
+const FlakyTestSchema = z.object({
+  flaky_tests: z.array(
+    z.object({
+      job_number: z.number(),
+    }),
+  ),
+  total_flaky_tests: z.number(),
+});
+
+const TestSchema = z.object({
+  message: z.string(),
+  run_time: z.union([z.string(), z.number()]),
+  file: z.string().optional(),
+  result: z.string(),
+  name: z.string(),
+  classname: z.string(),
+});
+
+const PaginatedTestResponseSchema = z.object({
+  items: z.array(TestSchema),
+  next_page_token: z.string().nullable(),
+});
+
+export const Test = TestSchema;
+export type Test = z.infer<typeof TestSchema>;
+
+export const PaginatedTestResponse = PaginatedTestResponseSchema;
+export type PaginatedTestResponse = z.infer<typeof PaginatedTestResponseSchema>;
+
+export const FlakyTest = FlakyTestSchema;
+export type FlakyTest = z.infer<typeof FlakyTestSchema>;
+
 // Export the schemas and inferred types with the same names as the original types
 export const Pipeline = PipelineSchema;
 export type Pipeline = z.infer<typeof PipelineSchema>;